source/requirements.txt
CherryPy==18.8.0
pyodbc>=5.0.0
openai>=1.3.0
bcrypt>=4.0.1
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
requests>=2.31.0
Jinja2>=3.1.2
